fix: Update mappings for research-object-type
Based on the discussion in #116.

Our `research-object-type` is a description of the content of the book /
chapters / OER but with a very specific list of possible values.

This metadata field / property does not describe the
nature/structure/datatype of the book, so mappings to type should not
be made here.

The mapping to MoDALIA shows that the content concerns itself with data
literacy (not just research data management) of a specific kind of data.